+++ label_studio/core/settings/base.py
@@ -432,6 +432,10 @@ SUPPORTED_EXTENSIONS = set(
     ]
 )
 
+# directory for files created during unit tests
+TEST_DATA_ROOT = os.path.join(BASE_DATA_DIR, 'test_data')
+os.makedirs(TEST_DATA_ROOT, exist_ok=True)
+
 # project exports
 EXPORT_DIR = os.path.join(BASE_DATA_DIR, 'export')
 EXPORT_URL_ROOT = '/export/'
@@ -504,6 +508,7 @@ IO_STORAGES_IMPORT_LINK_NAMES = [
 
 CREATE_ORGANIZATION = 'organizations.functions.create_organization'
 SAVE_USER = 'users.functions.save_user'
+POST_PROCESS_REIMPORT = 'core.utils.common.empty'
 USER_SERIALIZER = 'users.serializers.BaseUserSerializer'
 USER_SERIALIZER_UPDATE = 'users.serializers.BaseUserSerializerUpdate'
 TASK_SERIALIZER = 'tasks.serializers.BaseTaskSerializer'
@@ -575,6 +580,9 @@ FEATURE_FLAGS_FILE = get_env('FEATURE_FLAGS_FILE', 'feature_flags.json')
 FEATURE_FLAGS_OFFLINE = get_bool_env('FEATURE_FLAGS_OFFLINE', True)
 # default value for feature flags (if not overridden by environment or client)
 FEATURE_FLAGS_DEFAULT_VALUE = False
+
+# Whether to send analytics telemetry data
+COLLECT_ANALYTICS = get_bool_env('collect_analytics', True)
 
 # Strip harmful content from SVG files by default
 SVG_SECURITY_CLEANUP = get_bool_env('SVG_SECURITY_CLEANUP', False)